+++ Loyalify.Contracts/Offer/AddOfferRequest.cs
@@ -6,7 +6,7 @@ public class AddOfferRequest
 {
     public string Name { get; set; } = null!;
     public string Description { get; set; } = null!;
-    public int PointAmount { get; set; }
+    public decimal PointAmount { get; set; }
     public int StoreId { get; set; }
     public int ExpiresIn { get; set; }
     public IFormFile? ImageFile { get; set; }
