+++ Loyalify.Domain/Entities/Offer.cs
@@ -5,7 +5,7 @@ public class Offer
     public Guid Id { get; set; }
     public string Name { get; set; } = null!;
     public string Description { get; set; } = null!;
-    public int PointAmount { get; set; }
+    public decimal PointAmount { get; set; }
     public Store Store { get; set; } = null!;
     public DateTime CreatedAt { get; set; }
     public DateTime Deadline { get; set; }
